[Fix] Fix report model unit

+++ app/model/report.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 from typing import List, Optional
 from pydantic import BaseModel
 
@@ -5,6 +6,5 @@ class ReportBySurvey(BaseModel):
     question_id: int
     question_title: str
     survey_id: int
-    average_score: int
-    scorefactor_multiplied: int
-
+    average_score: Decimal
+    scorefactor_multiplied: Decimal
